feat: SAA를 위한 구조 변경
