+++ app/src/main/java/nbc/group/recipes/presentation/fragment/MainFragment.kt
@@ -1,0 +1,47 @@
+package nbc.group.recipes.presentation.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.fragment.findNavController
+import androidx.navigation.ui.setupWithNavController
+import nbc.group.recipes.R
+import nbc.group.recipes.databinding.FragmentMainBinding
+
+class MainFragment : Fragment() {
+
+    private var _binding: FragmentMainBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        _binding = FragmentMainBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        initNavigation()
+
+        binding.testButton.setOnClickListener {
+            findNavController().navigate(R.id.action_mainFragment_to_signInFragment)
+        }
+    }
+
+    private fun initNavigation() {
+        val navHostFragment = binding.fcv.getFragment<NavHostFragment>()
+        val navController = navHostFragment.navController
+        binding.bottomNavBar.setupWithNavController(navController)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
