+++ app/src/main/res/menu/menu.xml
@@ -1,26 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-
     <item
-        android:id="@+id/signInFragment3"
+        android:id="@+id/homeFragment"
         android:icon="@drawable/ic_home"
         android:title="home"/>
-
     <item
         android:id="@+id/mapFragment"
         android:icon="@drawable/ic_map"
         android:title="map"/>
-
     <item
         android:id="@+id/bookmarkFragment"
         android:icon="@drawable/ic_bookmark"
         android:title="bookmark"/>
-
     <item
         android:id="@+id/mypageFragment"
         android:icon="@drawable/ic_mypage"
         android:title="mypage"/>
-
-
-
 </menu>
