+++ app/src/main/res/values/strings.xml
@@ -16,6 +16,4 @@
     <string name="sign_in_fragment_sign_in_other_way_text">다른 계정으로 로그인</string>
 
 
-
-
 </resources>
